refactor: remove pc_compat and normalize path handling
- Remove pc_compat (-Z+) flag
  -> Implement automatic path separator normalization
- Update CMake minimum standard to newer version
- Remove unused .gitignore
- Minor fixes and cleanup
- Rename amx/float.c to amx/amxFloat.c
- Replace sprintf with snprintf in sc1.c, sc2.c, sc4.c, amxargs.c, amxcons.c, amxdgram.c, pawndbg.c
- 'rel' (release) support: `main() { rel name[5] = { "John" }; }`
- In Compiler Version v3.10.7, we allow empty statements: `main() { var x = 1; if (x) { ; } }`
- Add sizearray #pragma once
- Fix typos(commit: Daniel Cortez)

+++ source/compiler/memfile.h
@@ -26,7 +26,7 @@ typedef struct memfile_s
 memfile_t *memfile_creat(const char *name, size_t init);
 void memfile_destroy(memfile_t *mf);
 void memfile_seek(memfile_t *mf, long seek);
-int memfile_write(memfile_t *mf, const void *buffer, size_t size);
+size_t memfile_write(memfile_t *mf, const void *buffer, size_t size);
 size_t memfile_read(memfile_t *mf, void *buffer, size_t maxsize);
 long memfile_tell(const memfile_t *mf);
